codebert-base.commons-lang.mid-2246.idx-19006.2.mutant

+++ commons-lang/src/main/java/org/apache/commons/lang3/mutable/MutableFloat.java
@@ -254,7 +254,7 @@ public class MutableFloat extends Number implements Comparable<MutableFloat>, Mu
      * @since 3.5
      */
     public float addAndGet(final Number operand) {
-        this.value += operand.floatValue();
+        this.value += operand.doubleValue();
         return value;
     }
